refactor: relocate channel/color modules, drop constants.ts
