+++ examples/volume_rendering/main.ts
@@ -3,7 +3,7 @@ import GUI from "lil-gui";
 import { Idetik, OmeZarrImageSource, PerspectiveCamera, VolumeLayer } from "@";
 import { createExplorationPolicy } from "@/core/image_source_policy";
 import { OrbitControls } from "@/objects/cameras/orbit_controls";
-import type { ChannelProps } from "@/objects/textures/channel";
+import type { ChannelProps } from "@/core/channel";
 
 const url =
   "https://public.czbiohub.org/organelle_box/datasets/A549/organelle_box_crop_v1.zarr/CLTA/PFA/002000/";
