--- src/core/constants.ts
@@ -1,2 +1,0 @@
-// Maximum number of channels supported in multichannel images
-export const MAX_CHANNELS = 32;
