+++ test/helpers.ts
@@ -2,7 +2,7 @@ import type { Chunk } from "../src/data/chunk";
 import { Viewport } from "@/core/viewport";
 import { OrthographicCamera } from "@/objects/cameras/orthographic_camera";
 import { LayerManager } from "@/core/layer_manager";
-import { ChunkManager } from "@/core/chunk_manager";
+import { ChunkManager } from "@/data/chunk_manager";
 import type { IdetikContext } from "@/idetik";
 
 type ChunkOverrides = Partial<
